==> Rscripts/04_TSR_results.Rmd <==
---
title: "Temperature size rule results"
output: 
  html_document: 
    fig_caption: yes
    keep_md: yes
    toc: yes
    theme: "lumen"
    toc_float: true
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r message=FALSE, warning=FALSE, include=FALSE}
library(tidyverse)
library(broom)
library(gridExtra)
library(lubridate)
library(plotrix)
library(stringr)
```


### Background

Something here about the TSR being the third 'universal' response to warming. However, the mechanisms responsible for this widespread pattern are unknown blah blah blah, particularly w/r/t whether being smaller at warmer temperatures is associated with higher fitness. Or else why would this pattern be so prevalent in nature?


```{r message=FALSE, warning=FALSE}
data3 <- read_csv("/Users/Joey/Documents/Daph-TSR/data-processed/data3.csv")
```


### Reproductive rate
How does reproductive rate vary with temperature?
```{r reproductive rate, echo=FALSE, warning=FALSE}
data3 %>% 
	filter(unique_id != "K_16") %>% ## something weird is going on here!
	mutate(inverse_temp = (1/(.00008617*(temperature+273.15)))) %>%
	ggplot(data = ., aes(x = inverse_temp, y = log(time_to_first_clutch), label = id)) + geom_point(size = 4, color = "#619CFF") +
	geom_smooth(method = "lm", color = "#619CFF") +
	scale_x_reverse() + xlab("temperature (1/kT)") + ylab("ln time to reproductive maturity") + 
	theme_minimal() + 
	theme(axis.text.y   = element_text(size=20),
				axis.text.x   = element_text(size=20),
				axis.title.y  = element_text(size=20),
				axis.title.x  = element_text(size=20),
				panel.background = element_blank(),
				panel.grid.major = element_blank(), 
				panel.grid.minor = element_blank(),
				axis.line = element_line(colour = "black"),
				axis.ticks = element_line(size = 1)) +
	theme(panel.border = element_blank(), axis.line = element_line(colour="black", size=1, lineend="square"))
```

```{r}
data3 %>% 
	filter(unique_id != "K_16") %>% ## something weird is going on here!
	mutate(inverse_temp = (1/(.00008617*(temperature+273.15)))) %>%
	do(tidy(lm(log(time_to_first_clutch) ~ inverse_temp, data = .), conf.int = TRUE)) %>%
	knitr::kable(.)
```


Time between clutches

```{r time between clutches, echo=FALSE, warning=FALSE}
data3 %>% 
	filter(unique_id != "K_16") %>% ## something weird is going on here!
	mutate(inverse_temp = (1/(.00008617*(temperature+273.15)))) %>%
	ggplot(data = ., aes(x = inverse_temp, y = log(time_btw_1_2), label = id)) + geom_point(size = 4, color = "#619CFF") +
	geom_smooth(method = "lm", color = "#619CFF") +
	scale_x_reverse() + xlab("temperature (1/kT)") + ylab("time between clutches") +
	theme_minimal() + 
	theme(axis.text.y   = element_text(size=20),
				axis.text.x   = element_text(size=20),
				axis.title.y  = element_text(size=20),
				axis.title.x  = element_text(size=20),
				panel.background = element_blank(),
				panel.grid.major = element_blank(), 
				panel.grid.minor = element_blank(),
				axis.line = element_line(colour = "black"),
				axis.ticks = element_line(size = 1)) +
	theme(panel.border = element_blank(), axis.line = element_line(colour="black", size=1, lineend="square"))
```

Time to 4th clutch

```{r time between clutches, echo=FALSE, warning=FALSE}
data3 %>% 
	filter(unique_id != "K_16") %>% ## something weird is going on here!
	filter(temperature > 12) %>% 
	mutate(time_to_clutch4 = time_to_first_clutch + time_btw_1_2 + time_bwn_2_3 + time_bwn_3_4) %>% 
	group_by(temperature) %>% 
	ggplot(data = ., aes(x = length_at_1st_clutch, y = time_to_first_clutch, color = factor(temperature))) + geom_point(size = 4) +
	geom_smooth(method = "lm") +
 xlab("body size") + ylab("time to 1st clutch") +
	theme_minimal() + 
	theme(axis.text.y   = element_text(size=20),
				axis.text.x   = element_text(size=20),
				axis.title.y  = element_text(size=20),
				axis.title.x  = element_text(size=20),
				panel.background = element_blank(),
				panel.grid.major = element_blank(), 
				panel.grid.minor = element_blank(),
				axis.line = element_line(colour = "black"),
				axis.ticks = element_line(size = 1)) +
	theme(panel.border = element_blank(), axis.line = element_line(colour="black", size=1, lineend="square"))
```

```{r}
## time between clutches
data3 %>% 
	filter(unique_id != "K_16") %>% ## something weird is going on here!
	mutate(inverse_temp = (1/(.00008617*(temperature+273.15)))) %>%
	do(tidy(lm(log(time_btw_1_2) ~ inverse_temp, data = .), conf.int = TRUE)) %>%
	knitr::kable(.)
```

### Body size

```{r body size, echo=FALSE, warning = FALSE, message = FALSE}
data4 <- data3 %>% 
	gather(clutch_number, length, starts_with("length")) %>% 
	mutate(clutch_number = str_replace(clutch_number, "length_at_1st_clutch", "1st clutch")) %>% 
	mutate(clutch_number = str_replace(clutch_number, "length_at_2nd_clutch_um", "2nd clutch")) %>% 
	mutate(clutch_number = str_replace(clutch_number, "length_at_3rd_clutch", "3rd clutch")) %>% 
	mutate(clutch_number = str_replace(clutch_number, "length_at_4th_clutch", "4th clutch")) %>% 
	mutate(clutch_number = str_replace(clutch_number, "length_at_birth_um", "birth")) 
	

data4 %>% 
	filter(unique_id != "K_16") %>% ## something weird is going on here!
	mutate(inverse_temp = (1/(.00008617*(temperature+273.15)))) %>%
	group_by(clutch_number) %>% 
	ggplot(data = ., aes(x = inverse_temp, y = log(length), color = factor(clutch_number))) + geom_point(size = 5, alpha = 0.5) +
	geom_smooth(method = "lm") +
	scale_x_reverse() + xlab("temperature (1/kT)") + ylab("ln length (um)") +
	theme_minimal() + 
	theme(axis.text.y   = element_text(size=20),
				axis.text.x   = element_text(size=20),
				axis.title.y  = element_text(size=20),
				axis.title.x  = element_text(size=20),
				panel.background = element_blank(),
				panel.grid.major = element_blank(), 
				panel.grid.minor = element_blank(),
				axis.line = element_line(colour = "black"),
				axis.ticks = element_line(size = 1),
				legend.title = element_blank()) +
	theme(panel.border = element_blank(), axis.line = element_line(colour="black", size=1, lineend="square"))
```

### Somatic growth rates

```{r, echo=FALSE, message=FALSE, warning=FALSE}
## get to somatic growth rate by taking the difference in length at 1st clutch and length at birth, divided by the days to first clutch

data3 %>% 
	filter(unique_id != "K_16") %>% ## something weird is going on here!
	mutate(inverse_temp = (1/(.00008617*(temperature+273.15)))) %>%
	mutate(somatic_growth_rate = ((length_at_1st_clutch - length_at_birth_um)/time_to_first_clutch)) %>%
	ggplot(data = ., aes(x = inverse_temp, y = log(somatic_growth_rate), label = id)) + geom_point(size = 4, color = "#619CFF", alpha = 0.5) +
	geom_smooth(method = "lm", color = "#619CFF") +
	scale_x_reverse() + xlab("temperature (1/kT)") + ylab("ln somatic growth rate (um/day)") +
	theme_minimal() + 
	theme(axis.text.y   = element_text(size=20),
				axis.text.x   = element_text(size=20),
				axis.title.y  = element_text(size=20),
				axis.title.x  = element_text(size=20),
				panel.background = element_blank(),
				panel.grid.major = element_blank(), 
				panel.grid.minor = element_blank(),
				axis.line = element_line(colour = "black"),
				axis.ticks = element_line(size = 1),
				legend.title = element_blank()) +
	theme(panel.border = element_blank(), axis.line = element_line(colour="black", size=1, lineend="square"))
```

```{r}
data3 %>% 
	filter(unique_id != "K_16") %>% ## something weird is going on here!
	mutate(inverse_temp = (-1/(.00008617*(temperature+273.15)))) %>%
	mutate(somatic_growth_rate = ((length_at_1st_clutch - length_at_birth_um)/time_to_first_clutch)) %>%
do(tidy(lm(log(somatic_growth_rate) ~ inverse_temp, data = .), conf.int = TRUE)) %>%
	knitr::kable(.)
```

### Size rate trade-off??

```{r, echo=FALSE, message=FALSE, warning=FALSE}
data7 <- read_csv("/Users/Joey/Documents/Daph-TSR/data-processed/data7.csv")


data7 %>% 
	# filter(temperature > 13) %>% 
ggplot(aes(x = somatic_growth_rate, y = max_length, color = factor(temperature))) + geom_point(size = 4, alpha = 0.5) +
geom_smooth(method = "lm", color = "#619CFF") +
	xlab("somatic growth rate (um/day)") + ylab("max body length (um)") +
	theme_minimal() + 
	theme(axis.text.y   = element_text(size=20),
				axis.text.x   = element_text(size=20),
				axis.title.y  = element_text(size=20),
				axis.title.x  = element_text(size=20),
				panel.background = element_blank(),
				panel.grid.major = element_blank(), 
				panel.grid.minor = element_blank(),
				axis.line = element_line(colour = "black"),
				axis.ticks = element_line(size = 1),
				legend.title = element_blank()) +
	theme(panel.border = element_blank(), axis.line = element_line(colour="black", size=1, lineend="square"))
```

```{r}
data7 %>% 
# filter(clutch_number == "1st clutch") %>% 
	filter(temperature > 13) %>% 
		mutate(growth_rate_per_hour = somatic_growth_rate/24) %>% 
ggplot(aes(x = growth_rate_per_hour, y = max_length, color = factor(temperature))) + geom_point(size = 4, alpha = 0.5) +
geom_smooth(method = "lm", color = "#619CFF") +
	xlab("somatic growth rate (um/hour)") + ylab("max length (um)") +
	theme_minimal() 


data7 %>% 
	filter(temperature > 10) %>% 
	mutate(growth_rate_per_hour = somatic_growth_rate/24) %>% 
do(tidy(lm(log(max_length) ~ growth_rate_per_hour, data = .), conf.int = TRUE)) %>% View

```

```{r}
data3 %>% 
	filter(temperature > 10) %>% 
	mutate(time_to_mat = date_of_1st_clutch - date_of_birth_m_d_y) %>%
	mutate(time_to_mat = as.numeric(as.character(time_to_mat))) %>% 
	group_by(temperature) %>%
ggplot(aes(x =length_at_1st_clutch, y = 1/time_to_mat, color = factor(temperature))) + geom_point(size = 4, alpha = 0.5) +
geom_smooth(method = "lm") +
	facet_wrap( ~ temperature, scales = "free") +
	xlab("length at first clutch") + ylab("reproductive rate (1/time to maturity)") +
	theme_minimal() + 
	theme(axis.text.y   = element_text(size=10),
				axis.text.x   = element_text(size=10),
				axis.title.y  = element_text(size=12),
				axis.title.x  = element_text(size=12),
				panel.background = element_blank(),
				panel.grid.major = element_blank(), 
				panel.grid.minor = element_blank(),
				axis.line = element_line(colour = "black"),
				axis.ticks = element_line(size = 1),
				legend.title = element_blank()) +
	theme(panel.border = element_blank(), axis.line = element_line(colour="black", size=1, lineend="square"))
ggsave("figures/size_v_reproductive_rate.png", width = 6, height = 4)
ggsave("figures/first_clutch_length_v_reproductive_rate.png", width = 6, height = 4)
```

#### Body size over time
```{r, echo=FALSE, message=FALSE, warning=FALSE}
lengths_ann_raw <- read_csv("/Users/Joey/Documents/Daph-TSR/data-processed/lengths_all_adult_annotated.csv")
lengths_all <- read_csv("/Users/Joey/Documents/Daph-TSR/data-processed/lengths_all.csv")


lengths_all_clean <- lengths_all %>% 
	mutate(unique_id = str_replace(unique_id, "2N1.2", "N")) %>% 
	mutate(unique_id = str_replace(unique_id, "2N1.3", "N")) %>%
	mutate(unique_id = str_replace(unique_id, "2N1.4", "N")) %>% 
	mutate(unique_id = str_replace(unique_id, "2N1.8", "N")) %>%
	mutate(unique_id = str_replace(unique_id, "2N1.5", "N")) %>% 
	mutate(unique_id = str_replace(unique_id, "2N1.1", "N")) %>%
	separate(unique_id, into = c("letter", "temp"), remove = FALSE) %>%
	mutate(id = ifelse(is.na(id), ID, id)) %>% 
	mutate(unique_letter = str_extract(ID, "[A-Z]")) %>% 
	mutate(letter = ifelse(is.na(letter), unique_letter, letter)) %>%
	mutate(temp = ifelse(is.na(temp), temperature, temp)) 


lengths_all_clean %>% 
	filter(length > 1000) %>%
	select(temp, letter, date, length, version) %>% 
	filter(!is.na(temp)) %>% 
	group_by(temp) %>% 
	summarise_each(funs(mean, std.error, min), length) %>% 
	ggplot(aes(x = temp, y = mean)) + geom_point() +
	geom_errorbar(aes(ymin= mean - std.error, ymax = mean + std.error))


lengths_ann <- lengths_ann_raw %>% 
	mutate(date = mdy(date))

lengths_birth <- lengths_all %>%
	filter(life_stage == "birth") %>% 
	mutate(unique_id = str_replace(unique_id, "2N1.2", "N")) %>% 
	mutate(unique_id = str_replace(unique_id, "2N1.3", "N")) %>%
	mutate(unique_id = str_replace(unique_id, "2N1.4", "N")) %>% 
	mutate(unique_id = str_replace(unique_id, "2N1.8", "N")) %>%
	mutate(unique_id = str_replace(unique_id, "2N1.5", "N")) %>% 
	mutate(unique_id = str_replace(unique_id, "2N1.1", "N")) %>%
	separate(unique_id, into = c("letter", "temp"), remove = FALSE) 

lengths_b2m <- bind_rows(lengths_ann, lengths_birth)


### yes this is it!!
lengths_b2m %>% 
	mutate(certainty = ifelse(life_stage == "birth", "yes", certainty)) %>%
	# filter(version == 2) %>%
	filter(certainty == "yes") %>% 
	# filter(life_stage == "birth") %>% 
	# filter(temperature == 16) %>% 
	ggplot(aes(date, y = length, color = letter, group = letter)) + geom_point(size = 3) +
	geom_line() + facet_wrap( ~ temperature)
```


#### Minimum adult body size
```{r, echo=FALSE, message=FALSE, warning=FALSE}
lengths_all_adult <- read_csv("/Users/Joey/Documents/Daph-TSR/data-processed/lengths_all_adult.csv")

lengths_all_adult %>% 
filter(temperature > 10) %>% 
	mutate(inverse_temp = (1/(.00008617*(temperature+273.15)))) %>%
	filter(length > 1000) %>% 
	group_by(inverse_temp) %>% 
	summarise_each(funs(mean, min, max, std.error), length) %>%
	ggplot(aes(x = inverse_temp, y = min)) + geom_point(size = 4) +
	# geom_smooth() + 
	geom_errorbar(aes(ymin = min - std.error, ymax = min + std.error), width = 0.1) + 
	scale_x_reverse() + xlab("temperature (1/kT)") + ylab("minimum adult body length (um)") +
	theme_minimal() + 
	theme(axis.text.y   = element_text(size=20),
				axis.text.x   = element_text(size=20),
				axis.title.y  = element_text(size=20),
				axis.title.x  = element_text(size=20),
				panel.background = element_blank(),
				panel.grid.major = element_blank(), 
				panel.grid.minor = element_blank(),
				axis.line = element_line(colour = "black"),
				axis.ticks = element_line(size = 1),
				legend.title = element_blank()) +
	theme(panel.border = element_blank(), axis.line = element_line(colour="black", size=1, lineend="square"))
```


#### Maximum adult body size
```{r, echo=FALSE, message=FALSE, warning=FALSE}
lengths_all_adult %>% 
filter(temperature > 10) %>% 
	mutate(inverse_temp = (1/(.00008617*(temperature+273.15)))) %>%
	filter(length > 1000) %>% 
	group_by(inverse_temp) %>% 
	summarise_each(funs(mean, min, max, std.error), length) %>%
	ggplot(aes(x = inverse_temp, y = max)) + geom_point(size = 4) +
	# geom_smooth() + 
	geom_errorbar(aes(ymin = max - std.error, ymax = max + std.error), width = 0.1) + 
	scale_x_reverse() + xlab("temperature (1/kT)") + ylab("max adult body length (um)") +
	theme_minimal() + 
	theme(axis.text.y   = element_text(size=20),
				axis.text.x   = element_text(size=20),
				axis.title.y  = element_text(size=20),
				axis.title.x  = element_text(size=20),
				panel.background = element_blank(),
				panel.grid.major = element_blank(), 
				panel.grid.minor = element_blank(),
				axis.line = element_line(colour = "black"),
				axis.ticks = element_line(size = 1),
				legend.title = element_blank()) +
	theme(panel.border = element_blank(), axis.line = element_line(colour="black", size=1, lineend="square"))
```

#### Mean adult body size
```{r, echo=FALSE, message=FALSE, warning=FALSE}
lengths_all_adult %>% 
filter(temperature > 10) %>% 
	mutate(inverse_temp = (1/(.00008617*(temperature+273.15)))) %>%
	filter(length > 1000) %>% 
	group_by(inverse_temp) %>% 
	summarise_each(funs(mean, min, max, std.error), length) %>%
	ggplot(aes(x = inverse_temp, y = mean)) + geom_point(size = 4) +
	# geom_smooth() + 
	geom_errorbar(aes(ymin = mean - std.error, ymax = mean + std.error), width = 0.1) + 
	scale_x_reverse() + xlab("temperature (1/kT)") + ylab("average adult body length (um)") +
	theme_minimal() + 
	theme(axis.text.y   = element_text(size=20),
				axis.text.x   = element_text(size=20),
				axis.title.y  = element_text(size=20),
				axis.title.x  = element_text(size=20),
				panel.background = element_blank(),
				panel.grid.major = element_blank(), 
				panel.grid.minor = element_blank(),
				axis.line = element_line(colour = "black"),
				axis.ticks = element_line(size = 1),
				legend.title = element_blank()) +
	theme(panel.border = element_blank(), axis.line = element_line(colour="black", size=1, lineend="square"))
```

#### Next step: bring in the fecundity data to merge with the size data

```{r}

lengths_clean <- lengths_all_adult %>% 
	mutate(unique_id = str_replace(unique_id, "2N1.2", "N")) %>% 
	mutate(unique_id = str_replace(unique_id, "2N1.3", "N")) %>%
	mutate(unique_id = str_replace(unique_id, "2N1.4", "N")) %>% 
	mutate(unique_id = str_replace(unique_id, "2N1.8", "N")) %>%
	mutate(unique_id = str_replace(unique_id, "2N1.5", "N")) %>% 
	mutate(unique_id = str_replace(unique_id, "2N1.1", "N")) %>% 
	separate(unique_id, into = c("letter", "temp"), remove = FALSE) %>% 
	mutate(life_stage = ifelse(life_stage == "1", "clutch_1", life_stage)) %>% 
	mutate(life_stage = ifelse(life_stage == "2", "clutch_2", life_stage)) %>%
	mutate(life_stage = ifelse(life_stage == "3", "clutch_3", life_stage)) %>%
	mutate(life_stage = ifelse(life_stage == "4", "clutch_4", life_stage)) %>% 
	rename(clutch_number = life_stage)

	
data_raw <- read_csv("/Users/Joey/Documents/Daph-TSR/data-raw/DAPH-TSR-clutches.csv")

v2_babies <- data_raw %>% 
	filter(grepl("V2", ID)) %>% 
	separate(ID, into = c("V", "letter"), sep = 2) %>% 
	unite(unique_id, letter, temperature, sep = "_")


data_raw %>% 
	filter(str_detect(ID, "N")) %>% 
	arrange(temperature, clutch_number)### come back to filling out the clutch number business here, it's clearly not totally complete!


all <- left_join(lengths_clean, v2_babies, by = c("unique_id", "clutch_number"))

ggplot(data = all, aes(x = length, y = individuals, color = factor(temperature))) + geom_point(size = 4)


	
```	

For a given temperature, plot size vs. number of individuals

```{r}

all %>% 
filter(individuals != 0) %>% 
ggplot(data = ., aes(x = length, y = individuals, color = factor(temperature))) + geom_point(size = 4) + facet_wrap( ~ temperature)
```
How do we answer the question: is being smaller at a given temperature maximizing your fitness?
```{r}

all %>% 
	filter(individuals != 0) %>% 
	# filter(temperature == 24) %>%
ggplot(data = ., aes(x = length, y = individuals, color = factor(temperature))) + geom_point(size = 4) + geom_smooth(method = "lm") +
	facet_wrap( ~ temperature, scales = "free")
```
	
```{r}
all %>% 
	filter(temperature == 24) %>%
	filter(individuals != 0) %>% 
ggplot(data = ., aes(x = length, y = individuals)) + geom_point(aes(color = factor(letter)), size = 4) + geom_smooth(method = "lm") +
	facet_wrap( ~ temperature)
```


Ok so what I think we are seeing is that on a per individual clutch basis, there is still a positive relationship between body size and number of individuals per clutch...which I guess is not surprising...but maybe a more accurate measure of fitness is the number of offspring per unit time??? Maybe I should have looked at total lifetime reproductive output, not the number of offspring per clutch? It's clear that the warmer invidividuals pump through more clutches per unit time. Maybe I should look at whether within a temperature, you get through more clutches if you are smaller?

OK, so update: here's what I think I need to do...plot time to 4th clutch vs. size! Do you get through more clutches more quickly when you're smaller??


Maybe I should take the time to first reproduction and average clutch size and somehow plot that versus size at temperature?

```{r}

data3_select <- data3 %>% 
	select(id, temperature, time_to_first_clutch, length_at_1st_clutch, length_at_2nd_clutch_um) %>% 
	mutate(id = str_replace(id, "2N1.2", "N")) %>% 
	mutate(id = str_replace(id, "2N1.3", "N")) %>%
	mutate(id = str_replace(id, "2N1.1", "N")) %>%
	mutate(id = str_replace(id, "2N1.4", "N")) %>%
	mutate(id = str_replace(id, "2N1.5", "N"))
	
	


data3_select %>% 
	filter(time_to_first_clutch != 1176) %>% 
	filter(temperature > 13) %>% 
	ggplot(data = ., aes(x = time_to_first_clutch, y = length_at_1st_clutch, color = factor(temperature), label = id)) + geom_point(size = 4) + geom_smooth(method = "lm")

data3_select %>% 
	filter(time_to_first_clutch != 1176) %>% 
	filter(temperature > 13) %>% 
	ggplot(data = ., aes(x = time_to_first_clutch, y = length_at_1st_clutch)) + geom_point(aes(color = factor(temperature)), size = 4) + geom_smooth(method = "lm")

tidy(lm(length_at_1st_clutch ~ time_to_first_clutch, data = data3_select), conf.int = TRUE)
summary(lm(length_at_1st_clutch ~ time_to_first_clutch, data = data3_select))
```
